pkg/runner: move functions from cmd/lvh/runner to the pkg
This will be better to use those functions as library functions if they
are located in pkg/runner.
